Generalize deck backend architecture
  - Add generic productName and boardRevision fields to DeckInfo
  - Remove hardcoded TLV assumptions from core deck functions
  - Centralize driver finding in enumerateDecks() for all backends
  - Update OneWire and Forced backends to populate generic fields

  This enables non-TLV backends (like DeckCtrl) to work  with the existing deck system.
